Application: Move management of InAppAppealDialog to its container
- Also, move that widget to its own module because now it's independent
from InAppAppealStatus.
- And show in-app appeal on 5th and 25th time Spyder starts with a bit
of delay.
