+++ spyder/plugins/application/widgets/__init__.py
@@ -8,4 +8,5 @@
 """Widgets for the Application plugin."""
 
 from .about import AboutDialog  # noqa
+from .appeal import FakeInAppAppealDialog, InAppAppealDialog  # noqa
 from .status import InAppAppealStatus  # noqa
